MichelsonInteferometer.py
A Michelson interferometer was made famous when it confirmed the theory of relativity over the theory of absolute position, in an experiment by Michelson and Morley in 1887.

The device takes advantage of the nanoscopic scale at which lasers interact to measure tiny distances, or the slowing down of light as it passes through a particular material. This script was used to measure the refractive index of helium gas, although the constants used are specific to the exact set-up that I used.