Dissertation_Histograms.C
The script I used to interpret the simulation data and make plots of the dark matter detector the simulations were attempting to model as a series of useful histograms. This macro can also calculate data such as the mean and standard deviation of the variables calculated by the simulation.

This probably won't make much sense to anybody who isn't familiar with ROOT and particle physics...but this is probably the best demonstration of my complex data analysis and C++ coding skills I have, so I thought it was worth including.