Dissertation_PressureVersusTime.C
The other main script I used for data analysis in my dissertation project. This script evaluated how the signal amplitude of the simulated dark matter detector varied over time with ambient pressure.

This script is largely similar to the other one but also includes a pretty near demonstration of extracting data from massive (300k + lines and 10 columns) CSV files. My plan was to get it to plot the recorded temperature and pressure in the lab on the same axis, but unfortunately I got whatever the coding equivalent of "writer's block" is, and since deadlines were looming I gave up and commented out the bits that were meant to do this. The rest of it works beautifully, though!